requirements.txt
django-autoslug
celery==4.2.1
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.9.0
Pillow==5.3.0



